+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- [JUnit Platform Engine] Do not fail discovery on an empty rerun file with JUnit Platform 6.1.x ([#3216](https://github.com/cucumber/cucumber-jvm/pull/3216))
 
 ## [7.34.5] - 2026-07-23
 ### Fixed